Skip database creation instead of running outside of the Django runner.
Makes it easier to configure tests on a Django project.

(cherry picked from commit e495f62)
