+++ examples/users/users/settings_mocked.py
@@ -1,0 +1,5 @@
+from django_mock_queries.utils import monkey_patch_test_db
+
+from users.settings import *
+
+monkey_patch_test_db()
