+++ tox.ini
@@ -10,5 +10,5 @@ flake8-max-line-length = 120
 deps = -rrequirements/requirements-testing.txt
 commands =
     py.test django_mock_queries/ tests/ --cov-report term-missing --cov=django_mock_queries --flake8
-    python -m unittest discover -s examples/users/ -p 'tests_mock*'
+    python -c "import subprocess; subprocess.check_call(['./manage.py', 'test', '--settings=users.settings_mocked'], cwd='examples/users')"
     python -c "import subprocess; subprocess.check_call(['./manage.py', 'test'], cwd='examples/users')"
